Ensures drush is avaiable and mutagen syncs
Drush was erroring as unavailable and this seems to be from a delay in
mutagen syncing. Considered checking mutagen status but that also seems
to have a delay. Rather than place a sleep command generally we loop to
see when Drush becomes available and timeout if it does not become
available